refactor: split problem CLI modules
